Merge upstream (#28)
* Make profile page responsive (#366)

* 374 Fix active reservations boundary logic (#375)

* Update filters to show active reservations on ambassador view

* Fix bug when start_idx < 0

* Only consider rooms that are present in map

---------

Co-authored-by: Yuvraj Jain <yjain@unc.edu>

* Reset policies for Phase 2 Launch of XL (#376)

* Fix broken build in prod due to npm memory limit (#377)

* Fix @angular-devkit/build-angular dependency vulnerabilities in production (#378)

* Fix vulnerability with @angular-devkit/build-angular dep

* Try reducing memory to match cloudapps pod limit (#379)

* Increase memory limit for deployment (#381)

* Hot fixing office hours schedule (#382)

* Fix bug with end_idx logic for future reservations (#384)

* Fix bug with end_idx logic for future reservations

The current problem with the code is that at 11 pm, we were
not able to see future reservations in red for tomorrow. This was
actually a red herring that led me to look at parts of the code
that were not responsible for this.

Then suddenly at 12 am, everything started working as intended.
This was the biggest hint for me in debugging this.

Basically on line 345, I just had to add in a condition to check
that the date I am getting as a param is the same date as today.
Only in that case, should I execute the logic of
end_idx < current_time_idx. However, earlier this was being
executed for all reservations which is why whenever we had
any reservations which had an end time of less than 11 pm
(which most reservations do), our function would just skip over
them. This also aligns with what we saw at 12 am when
everything seemed to work as intended. This is again because
most reservations don't have an ending time before 12 am.

Anyway, took me over 60 minutes to add in these 30 chars to fix
the bug. So I typed an average of 0.5 char per minute... :,)

* Added kind of a test for future reservations

Although this test is time dependent, and
we will need some way of freezing time
to accurately test the efficacy of the code.

* Change Dockerfile to use yarn instead of npm (#393)

* 399 add checkin time to avoid cancellation verbiage (#400)

* Add checkin deadline verbiage

* Add parameter to checkinDeadline

Since users can reserve a reservation up until the end a reservation time period, the checkinDeadline was exceeding the reservation's end time.
To correct this take the minimum of the start time + 10 mins and the end time.

---------

Co-authored-by: Jade Keegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: aarjav <97541934+aarjavjain2002@users.noreply.github.com>
Co-authored-by: Yuvraj Jain <yjain@unc.edu>
Co-authored-by: Kris Jordan <kris@cs.unc.edu>
Co-authored-by: Yuvraj Jain <97535654+yuvrajjain2003@users.noreply.github.com>
Co-authored-by: Noah Smith <nsmith2@unc.edu>
Co-authored-by: John Schachte <89204366+JohnSchachte@users.noreply.github.com>